[📚Docs] UseCase가 사용되는 책임을 Interface로 분리
