+++ PalangPalang/Sources/Domain/UseCase/AlarmUseCase.swift
@@ -8,29 +8,11 @@
 
 import Foundation
 
-protocol AlarmSettings {
-  func addAlarm(alarm: AlarmModel)
-}
-
-protocol AlarmOnProcess {
-  func readMissionDate() -> Date?
-}
-
-protocol MissionOnProcess {
-  func readTimeoutDate() -> Date?
-  func deleteAlarm()
-}
-
-protocol CheckAlarmStatus {
-  func verifyAndChangeAlarmState()
-}
-
 @Observable
-class AlarmUseCase: CheckAlarmStatus {
+class AlarmUseCase: AlarmStaus {
   static let shared = AlarmUseCase()
   
-  
-  var alarmState: AlarmState = .alarmOnSettings
+  var alarmStatus: AlarmState = .main
   private var timerService: TimerSchedulerService?
   
   private init() {
@@ -43,11 +25,8 @@ class AlarmUseCase: CheckAlarmStatus {
     timerService = TimerSchedulerService.init(
       alarm: alarm,
       onMissionStart: { [weak self] in
-        print("미션 시작 시 불리는 메서드")
         self?.verifyAndChangeAlarmState()
       }, onMissionFinish: { [weak self] in
-        print("미션 종료 시 불리는 메서드 ")
-//        print("미션 TimeOver")
         self?.verifyAndChangeAlarmState()
       }
     )
@@ -56,7 +35,7 @@ class AlarmUseCase: CheckAlarmStatus {
   /// 상태 갱신 요청
   func verifyAndChangeAlarmState() {
     let nowAlarm = DataStoreService<AlarmModel>().load()
-    self.alarmState = determineAppState(alarm: nowAlarm)
+    self.alarmStatus = determineAppState(alarm: nowAlarm)
   }
   
   /// Alarm 값을 토대로 Alarm 상태 분석
@@ -64,7 +43,7 @@ class AlarmUseCase: CheckAlarmStatus {
     let currentDate = Date()
     
     // AlarmModel이 없는 경우
-    guard let alarm = alarm else { return .alarmOnSettings }
+    guard let alarm = alarm else { return .main }
 
     // AlarmModel이 있는 경우
     if currentDate < alarm.dueDate {
@@ -77,7 +56,7 @@ class AlarmUseCase: CheckAlarmStatus {
   }
 }
 
-extension AlarmUseCase: AlarmSettings {
+extension AlarmUseCase: MainAlarmSettings {
   func addAlarm(alarm: AlarmModel) {
     DataStoreService<AlarmModel>.init().save(alarm)
     LocalNotificationService.init().scheduleNotification(title: "미션시작", body: "귀신 퇴치하러 갑시당!", at: alarm.dueDate)
@@ -87,9 +66,9 @@ extension AlarmUseCase: AlarmSettings {
 }
 
 extension AlarmUseCase: AlarmOnProcess {
-  func readMissionDate() -> Date? {
-    if let dueDate = DataStoreService<AlarmModel>.init().load()?.dueDate {
-      return dueDate
+  func readAlarmDate() -> (start: Date, due: Date)? {
+    if let alarm = DataStoreService<AlarmModel>.init().load() {
+      return (alarm.startDate, alarm.dueDate)
     } else {
       verifyAndChangeAlarmState() // 데이터가 없다면 상태 갱신 요청
       return nil
