+++ PalangPalang/Sources/Domain/UseCase/Interface/AlarmStaus.swift
@@ -1,0 +1,13 @@
+//
+//  AlarmStaus.swift
+//  PalangPalang
+//
+//  Created by 박혜운 on 8/2/24.
+//  Copyright © 2024 com.mc3. All rights reserved.
+//
+
+import Foundation
+
+protocol AlarmStaus {
+  var alarmStatus: AlarmState {get set}
+}
