+++ PalangPalang/Sources/Domain/UseCase/Interface/MainAlarmSettings.swift
@@ -1,0 +1,13 @@
+//
+//  MainAlarmSettings.swift
+//  PalangPalang
+//
+//  Created by 박혜운 on 8/2/24.
+//  Copyright © 2024 com.mc3. All rights reserved.
+//
+
+import Foundation
+
+protocol MainAlarmSettings {
+  func addAlarm(alarm: AlarmModel)
+}
