+++ PalangPalang/Sources/Domain/UseCase/Interface/MissionOnProcess.swift
@@ -1,0 +1,14 @@
+//
+//  MissionOnProcess.swift
+//  PalangPalang
+//
+//  Created by 박혜운 on 8/2/24.
+//  Copyright © 2024 com.mc3. All rights reserved.
+//
+
+import Foundation
+
+protocol MissionOnProcess {
+  func readTimeoutDate() -> Date?
+  func deleteAlarm()
+}
